move tooltips
