+++ containers/ecr-viewer/src/app/view-data/components/EcrMetadata.tsx
@@ -4,7 +4,7 @@ import {
   AccordionDiv,
 } from "../component-utils";
 import { Table } from "@trussworks/react-uswds";
-import { toolTipElement } from "@/app/utils";
+import { ToolTipElement } from "@/app/ToolTipElement";
 import { ReportableConditions } from "../../services/ecrMetadataService";
 import { DataDisplay, DisplayDataProps } from "@/app/DataDisplay";
 import React from "react";
@@ -79,22 +79,28 @@ const EcrMetadata = ({
           <thead>
             <tr>
               <th className="reportability_summary_header">
-                {toolTipElement(
-                  "Reportable Condition",
-                  "List of conditions that caused this eCR to be sent to your jurisdiction based on the rules set up for routing eCRs by your jurisdiction in RCKMS (Reportable Condition Knowledge Management System). Can include multiple Reportable Conditions for one eCR.",
-                )}
+                <ToolTipElement
+                  content={"Reportable Condition"}
+                  toolTip={
+                    "List of conditions that caused this eCR to be sent to your jurisdiction based on the rules set up for routing eCRs by your jurisdiction in RCKMS (Reportable Condition Knowledge Management System). Can include multiple Reportable Conditions for one eCR."
+                  }
+                />
               </th>
               <th>
-                {toolTipElement(
-                  "RCKMS Rule Summary",
-                  "Reason(s) that this eCR was sent for this condition. Corresponds to your jurisdiction's rules for routing eCRs in RCKMS (Reportable Condition Knowledge Management System).",
-                )}
+                <ToolTipElement
+                  content={"RCKMS Rule Summary"}
+                  toolTip={
+                    "Reason(s) that this eCR was sent for this condition. Corresponds to your jurisdiction's rules for routing eCRs in RCKMS (Reportable Condition Knowledge Management System)."
+                  }
+                />
               </th>
               <th className="reportability_summary_header">
-                {toolTipElement(
-                  "Jurisdiction Sent eCR",
-                  "List of jurisdictions this eCR was sent to. Can include multiple jurisdictions depending on provider location, patient address, and jurisdictions onboarded to eCR.",
-                )}
+                <ToolTipElement
+                  content={"Jurisdiction Sent eCR"}
+                  toolTip={
+                    "List of jurisdictions this eCR was sent to. Can include multiple jurisdictions depending on provider location, patient address, and jurisdictions onboarded to eCR."
+                  }
+                />
               </th>
             </tr>
           </thead>
